Move adaptation options code to separate file
