+++ containers/web-server/frontend/src/sidebar/sections/networks/NetworkControl.tsx
@@ -8,12 +8,12 @@ import { CheckboxTree } from '@/lib/controls/checkbox-tree/CheckboxTree';
 import { NETWORK_LAYERS_HIERARCHY } from '@/config/networks/hierarchy';
 import { NETWORKS_METADATA } from '@/config/networks/metadata';
 import { LayerLabel } from '@/sidebar/ui/LayerLabel';
+import { showAdaptationsState } from '@/state/data-selection/networks/adaptations';
 import {
   networkTreeCheckboxState,
   networkTreeConfig,
   networkTreeExpandedState,
 } from '@/state/data-selection/networks/network-selection';
-import { showAdaptationsState } from '@/state/layers/data-layers/networks';
 
 export const NetworkControl: FC<{}> = () => {
   const [checkboxState, setCheckboxState] = useRecoilState(networkTreeCheckboxState);
